requirements.txt
numpy
pandas
python-dotenv
mysql-connector-python
pymysql
scikit-learn


# It automatically trigger and execute setup.py witout executing manually pyton setup.py install 
# -e .